Add files via upload
This project is a Snake and Ladder game implemented in C. It’s a classic board game where players race to reach the end of the board by rolling dice and navigating through various obstacles and shortcuts. The game features both snakes and ladders that dynamically affect the player's progress. Here I have done using C language with the help of arrays and functions. Here I have used the time.h function to get the random dice rolls. The gave is without any visual representation.